RandomUtils rename for clarity
